Separated cart expenditure operations
