+++ src/main/java/es/cloudapps/hexarch/infrastructure/api/CartExpenditureRestApiController.java
@@ -1,7 +1,7 @@
 package es.cloudapps.hexarch.infrastructure.api;
 
-import es.cloudapps.hexarch.hexagon.application.ShoppingCartServicePort;
-import es.cloudapps.hexarch.hexagon.application.ShoppingCartServicePort.*;
+import es.cloudapps.hexarch.hexagon.application.CartExpenditureServicePort;
+import es.cloudapps.hexarch.hexagon.application.CartExpenditureServicePort.*;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -10,15 +10,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api")
 public class CartExpenditureRestApiController {
 
-    private final ShoppingCartServicePort shoppingCartService;
+    private final CartExpenditureServicePort cartExpenditureService;
 
-    public CartExpenditureRestApiController(ShoppingCartServicePort shoppingCartService) {
-        this.shoppingCartService = shoppingCartService;
+    public CartExpenditureRestApiController(CartExpenditureServicePort cartExpenditureService) {
+        this.cartExpenditureService = cartExpenditureService;
     }
 
     @GetMapping("/cartexpenditure")
     public CartExpenditureResp getCartExpenditure() {
-        return shoppingCartService.getCartExpenditure(new CartExpenditureReq());
+        return cartExpenditureService.getCartExpenditure(new CartExpenditureReq());
     }
 
 }
